Optimise some code to reduce memory requirements. This involves inlining a few select PROCs (to avoid the overhead of a CALL; 4 words) and optimisation of pin.location to use slightly less memory.
  * pin.location (...)
    - now uses three arrays of port addresses to translate the given pin
      number into the port addresses, avoiding the previous parallel assignment
      of r.ddr, r.port, r.pin, and bit, saving 2 words.
    - inline the function so that it does not incur a cost of 4 words for the
      function call
    - pin.location now also checks if the pin is < 0 and ends the program with
      die if this is the case
  * die (...)
    - inline to avoid the cost of 4 words for the function call (which we would
      not expect to need very often as die signifies an error condition and
      stops the VM; programs that call die a lot are not very useful)
  * wait.for.interrupt
    - inline, as this is solely a wrapper function to hide the FFI name

These changes take the current blink.occ (which sets up 18 individual processes
in a replicated PAR, each one blinking one pin on the board) from using 583
words to using 331 words. This is at the cost of some code expansion (404 to 560
bytes).

The problem was that the accumulated cost of running each flash.pin process was
very large due to the memory requirements of digitalWrite (22 words). Inlining
die and pin.location reduces the amount of memory that has to be allocated due
to the possibility of the call sequence: digitalWrite -> pin.location -> die
happening (needing at least 12 words). Die should almost never happen and
pin.location can be be inlined to avoid a further 4 words being lost on a call.
Together with the 2 words saved on optimising pin.location this reduces
digitalWrite's memory requirements from 22 to 8 words.