Modified plumbing.module's PROC adc to allow for the ATmega1280 to read from all 16 of its adc pins.
This is definitely architecture specific code. Might even be board specific. Not sure about that one. The Arduino Mega and the Ardupilot Mega use the same layout for adc, I guess we'll see what's what when the Seeeduinos arive.
Perhaps we should find a better name for pinmapping.module, and include in that all the board-specific pin-goodness? I'm willing to bet pwm and serial code is going to have to be pulled out of plumbing.module just as adc is.