I flipped things over to the 3.3V configuration. This should, I think, run the 1280 at 8MHz. (I actually now need to go check.)
Actually, I assumed that. I'm not at all certain what speed the board runs at when you select 3.3V. It might still run at 16MHz. 

Still, I need to commit a pinmapping module if things are going to work. Note that a seeedmega33 isn't actually different from a seeedmega... at least, not as far as pin mappings go. For that reason, it might be that I don't need another platform directory for that configuration. (We might need a different .conf for the lower voltage configuration, but not a different platform.)