Servos are working for the Arduino and the Mega complies, dispite not having any real PWM or Servo support. This is probably the last big commit (from me, at least), before Matt and I start tearing down the tree (again) and making it better (but first broken).
I added a usec tag (for micro-seconds. Perhaps musec is a better name?) to the SERVO PROTOCOL, for use if we want to directly control the phase length on a servo, and so that pos only keeps track of where in the divs we are. setting pos now begins a recalculation of usec, which is written to pwm#.something (st, uff). Still need to add sanity checks and such to servo (..), but that can come later.