A good mapping for the Arduino board.
I believe this mapping is sound. A more comprehensive set of tests would be great. 

Note that everything in this module *must* be in terms of "board pins." The analog pins must, therefore, carry a constant value that indexes into the map, which then returns the AVR pin that they correspond to. All of our higher level code should be using "board.to.chip", and that in turn can use things like "chip.to.gpio" from the Arch directory...