Adding decimal printing.
I have an old, not-at-all optimal decimal printing routine. It might as well go in the printing library, and I might be able to circle around and optimize it later. At worst, it's there and get used.

twi2.module will go away, ultimately, as I will bring in the TWI code that I'm evolving with the ES sensor project. So, this can commit for now, but it will go away eventually.